rm __pycach__
