--- pocApp/apps.py
@@ -1,13 +1,0 @@
-"""
-.. _apps:
-django apps module for NMap app
-:copyright:
-    Copyright 2018 Provincial Health Service Authority
-    of British Columbia
-:contact:    ali.rahmat@phsa.ca
-"""
-__updated__ = '2018_08_21'
-from django.apps import AppConfig
-
-class PocappConfig(AppConfig):
-    name = 'pocApp'
