+++ ssl_cert_tracker/views.py
@@ -1,3 +1,22 @@
-from django.shortcuts import render
+"""
+django models for the ssl_certificates app
 
-# Create your views here.
+:module:    ssl_certificates.models
+
+:copyright:
+
+    Copyright 2018 Provincial Health Service Authority
+    of British Columbia
+
+:contact:    ali.rahmat@phsa.ca
+
+"""
+
+from django.http import HttpResponse
+
+
+#import ipdb;ipdb.set_trace()
+
+def index(request):
+    return HttpResponse('The Call is Success' + str(request))
+
